add comment id to comment table to replay, drop comment replay table and fix some services

+++ SocialMedia.Service/BlockService/BlockService.cs
@@ -6,6 +6,7 @@ using SocialMedia.Data.Models;
 using SocialMedia.Data.Models.ApiResponseModel;
 using SocialMedia.Data.Models.Authentication;
 using SocialMedia.Repository.BlockRepository;
+using SocialMedia.Repository.FollowerRepository;
 using SocialMedia.Repository.FriendRequestRepository;
 using SocialMedia.Repository.FriendsRepository;
 using SocialMedia.Service.GenericReturn;
@@ -17,14 +18,17 @@ namespace SocialMedia.Service.BlockService
         private readonly IBlockRepository _blockRepository;
         private readonly IFriendsRepository _friendsRepository;
         private readonly IFriendRequestRepository _friendRequestRepository;
+        private readonly IFollowerRepository _followerRepository;
         private readonly UserManagerReturn _userManagerReturn;
         public BlockService(IBlockRepository _blockRepository, IFriendsRepository _friendsRepository,
-            IFriendRequestRepository _friendRequestRepository, UserManagerReturn _userManagerReturn)
+            IFriendRequestRepository _friendRequestRepository, UserManagerReturn _userManagerReturn,
+            IFollowerRepository _followerRepository)
         {
             this._blockRepository = _blockRepository;
             this._friendsRepository = _friendsRepository;
             this._friendRequestRepository = _friendRequestRepository;
             this._userManagerReturn = _userManagerReturn;
+            this._followerRepository = _followerRepository;
         }
         public async Task<ApiResponse<Block>> BlockUserAsync(AddBlockDto addBlockDto, SiteUser user)
         {
@@ -32,30 +36,21 @@ namespace SocialMedia.Service.BlockService
                     addBlockDto.UserIdOrUserNameOrEmail);
             if (blockedUser != null)
             {
+                addBlockDto.UserIdOrUserNameOrEmail = blockedUser.Id;
                 var isUserBlocked = await _blockRepository.GetBlockByUserIdAndBlockedUserIdAsync(
                         user.Id, addBlockDto.UserIdOrUserNameOrEmail);
                 if (isUserBlocked == null)
                 {
-                    addBlockDto.UserIdOrUserNameOrEmail = blockedUser.Id;
-                    var isUserFriend = await _friendsRepository.GetFriendByUserAndFriendIdAsync(
-                            user.Id, addBlockDto.UserIdOrUserNameOrEmail);
-                    if (isUserFriend != null)
-                    {
-                        return await DeleteFromFriendListAndBlockAsync(addBlockDto, user);
-                    }
-                    var friendRequest = await _friendRequestRepository.GetFriendRequestByUserAndPersonIdAsync(
-                    user.Id, addBlockDto.UserIdOrUserNameOrEmail);
-                    if (friendRequest != null)
-                    {
-                        return await RemoveFriendRequestAndBlockAsync(addBlockDto, friendRequest, user);
-                    }
+                    await UnfollowAsync(addBlockDto, user);
+                    await DeleteFriendRequestAsync(addBlockDto, user);
+                    await DeleteFromFriendListAsync(addBlockDto, user);
                     var newBlockedUser = await _blockRepository.BlockUserAsync(
                             ConvertFromDto.ConvertFromBlockDto_Add(addBlockDto, user));
                     return StatusCodeReturn<Block>
                     ._200_Success("User blocked successfully", newBlockedUser);
                 }
                 return StatusCodeReturn<Block>
-                    ._403_Forbidden();
+                    ._403_Forbidden("User already blocked");
             }
             return StatusCodeReturn<Block>
                         ._404_NotFound("User you want to block not found");
@@ -67,20 +62,20 @@ namespace SocialMedia.Service.BlockService
             if (block == null)
             {
                 return StatusCodeReturn<Block>
-                    ._404_NotFound("This block");
+                    ._404_NotFound("Block not found");
             }
             return StatusCodeReturn<Block>
                 ._200_Success("Block found successfully", block);
         }
 
-        public async Task<ApiResponse<Block>> GetBlockByUserIdAndBlockedUserIdAsync
-            (string userId, string blockedUserId)
+        public async Task<ApiResponse<Block>> GetBlockByUserIdAndBlockedUserIdAsync(string userId,
+            string blockedUserId)
         {
             var block = await _blockRepository.GetBlockByUserIdAndBlockedUserIdAsync(userId, blockedUserId);
             if (block == null)
             {
                 return StatusCodeReturn<Block>
-                    ._400_BadRequest("User is not in your friend list");
+                    ._404_NotFound("User is not in your block list");
             }
             return StatusCodeReturn<Block>
                 ._200_Success("Blocked user found successfully", block);
@@ -110,7 +105,7 @@ namespace SocialMedia.Service.BlockService
                     ._200_Success("Blocked users found successfully", userBlockList);
         }
 
-        public async Task<ApiResponse<Block>> UnBlockUserAsync(UpdateBlockDto updateBlockDto, SiteUser user)
+        public async Task<ApiResponse<Block>> UnBlockUserAsync(UnBlockDto updateBlockDto, SiteUser user)
         {
             var blockedUser = await _userManagerReturn.GetUserByUserNameOrEmailOrIdAsync(
                 updateBlockDto.UserIdOrUserNameOrEmail);
@@ -124,45 +119,72 @@ namespace SocialMedia.Service.BlockService
                     var unblockedUser = await _blockRepository.UnBlockUserAsync(
                             ConvertFromDto.ConvertFromBlockDto_Update(updateBlockDto, user));
                     return StatusCodeReturn<Block>
-                            ._200_Success("User unblocked successfully", unblockedUser);
+                            ._200_Success("Block removed successfully", unblockedUser);
                 }
                 return StatusCodeReturn<Block>
-                        ._403_Forbidden();
+                        ._404_NotFound("User not in your block list");
             }
             return StatusCodeReturn<Block>
                         ._404_NotFound("User you want to unblock not found");
         }
 
-        public async Task<ApiResponse<Block>> UnBlockUserByBlockIdAsync(string blockId)
+        public async Task<ApiResponse<Block>> UnBlockUserByBlockIdAsync(string blockId, SiteUser user)
         {
             var block = await _blockRepository.GetBlockByIdAsync(blockId);
             if (block != null)
             {
-                var unblock = await _blockRepository.UnBlockUserAsync(block);
-                return StatusCodeReturn<Block>
-                ._200_Success("Unblocked successfully");
+                if(block.UserId == user.Id)
+                {
+                    var unblock = await _blockRepository.UnBlockUserAsync(block);
+                    return StatusCodeReturn<Block>
+                    ._200_Success("Unblocked successfully", unblock);
+                }
+                return StatusCodeReturn<Block>._403_Forbidden();
             }
             return StatusCodeReturn<Block>
                     ._404_NotFound("Block not found");
         }
 
-        private async Task<ApiResponse<Block>> DeleteFromFriendListAndBlockAsync(
+        private async Task<ApiResponse<Block>> DeleteFromFriendListAsync(
             AddBlockDto addBlockDto, SiteUser user)
         {
-            await _friendsRepository.DeleteFriendAsync(user.Id, addBlockDto.UserIdOrUserNameOrEmail);
-            var blockedUser = await _blockRepository.BlockUserAsync(
-            ConvertFromDto.ConvertFromBlockDto_Add(addBlockDto, user));
-            return StatusCodeReturn<Block>
-                ._200_Success("User blocked successfully", blockedUser);
+            var isUserFriend = await _friendsRepository.GetFriendByUserAndFriendIdAsync(
+                            user.Id, addBlockDto.UserIdOrUserNameOrEmail);
+            if (isUserFriend != null)
+            {
+                await _friendsRepository.DeleteFriendAsync(user.Id, addBlockDto.UserIdOrUserNameOrEmail);
+                return StatusCodeReturn<Block>._200_Success("Success");
+            }
+            return StatusCodeReturn<Block>._404_NotFound("You are not friends");
         }
-        private async Task<ApiResponse<Block>> RemoveFriendRequestAndBlockAsync(
-            AddBlockDto addBlockDto, FriendRequest friendRequest, SiteUser user)
+        private async Task<ApiResponse<Block>> DeleteFriendRequestAsync(
+            AddBlockDto addBlockDto, SiteUser user)
         {
-            await _friendRequestRepository.DeleteFriendRequestByAsync(friendRequest.Id);
-            var blockedUser = await _blockRepository.BlockUserAsync(
-            ConvertFromDto.ConvertFromBlockDto_Add(addBlockDto, user));
-            return StatusCodeReturn<Block>
-            ._200_Success("User blocked successfully", blockedUser);
+            var friendRequest = await _friendRequestRepository.GetFriendRequestByUserAndPersonIdAsync(
+                    user.Id, addBlockDto.UserIdOrUserNameOrEmail);
+            if (friendRequest != null)
+            {
+                await _friendRequestRepository.DeleteFriendRequestByAsync(friendRequest.Id);
+                return StatusCodeReturn<Block>._200_Success("Success");
+            }
+            return StatusCodeReturn<Block>._404_NotFound("Friend request not found");
+        }
+
+        private async Task<ApiResponse<Block>> UnfollowAsync(AddBlockDto addBlockDto, SiteUser user)
+        {
+            var isFollowingYou = await _followerRepository.GetFollowingByUserIdAndFollowerIdAsync(
+                    user.Id, addBlockDto.UserIdOrUserNameOrEmail);
+            var areYouFollowingHim = await _followerRepository.GetFollowingByUserIdAndFollowerIdAsync(
+                addBlockDto.UserIdOrUserNameOrEmail, user.Id);
+            if (isFollowingYou != null)
+            {
+                await _followerRepository.UnfollowAsync(user.Id, addBlockDto.UserIdOrUserNameOrEmail);
+            }
+            if (areYouFollowingHim != null)
+            {
+                await _followerRepository.UnfollowAsync(addBlockDto.UserIdOrUserNameOrEmail, user.Id);
+            }
+            return StatusCodeReturn<Block>._200_Success("Success");
         }
